requirements.txt
faiss-cpu
sentence-transformers
pandas
numpy
ollama
